Update Vagrant setup to bento/ubuntu-18.04 with PHP 7.2

+++ vagrant-provision.sh
@@ -5,23 +5,15 @@ export DEBIAN_FRONTEND=noninteractive
 set -e
 
 apt-get update
-apt-get -y upgrade
+apt-get install -y php curl php-mysql mysql-server php-xml php-mbstring php-zip unzip
 
-apt-get install -y php5 curl php5-mysql mysql-server uuid uuid-dev build-essential php-pear
-
-curl -sS https://getcomposer.org/installer | php -- --install-dir=/usr/local/bin
-chmod +x /usr/local/bin/composer.phar
-ln -s /usr/local/bin/composer.phar /usr/local/bin/composer
-
-cd /usr/local/bin
-wget https://phar.phpunit.de/phpunit.phar
-chmod +x phpunit.phar
-ln -s phpunit.phar phpunit
-
-printf "\n" | pecl install uuid
-echo 'extension=uuid.so' > /etc/php5/conf.d/uuid.ini
+if [ ! -f /usr/local/bin/composer ]; then
+  curl -sS https://getcomposer.org/installer | php -- --install-dir=/usr/local/bin
+  chmod +x /usr/local/bin/composer.phar
+  ln -s /usr/local/bin/composer.phar /usr/local/bin/composer
+fi
 
 cd /vagrant
-composer install
+su -c "cd /vagrant && composer install" vagrant
 
 mysql -uroot -e 'CREATE DATABASE IF NOT EXISTS emphloyer_test'
